Convert getChannel to composition of functions
I only noticed this problem after importing into a shell, which I could
only do once the Maya imports were wrapped in try/excepts. Problem: The
partially-applied getChannelAtTime function, which comprises the body of
the getChannel function, is evaluated strictly by Python, but itself
calls cmds.currentTime, which, of course, doesn't exist when this
library is loaded in a shell.