Add pure.lerp with tests
I went with the t arg first. I can see reasons to put it last, as I've
done in another library (functional images stuff), but it's convenient
to have it first with a current need, and that's how Edward Kmett does
it in his Linear library, and who am I to argue with Kmett?