Add tf.xyzDiv, tests
This is different than the other xyz<ArithOp> functions, as division can
result in non-integers. This is something Haskell deals with well, and I
had the goodness of element type homogeneity in container types driven
home yet again when I tried to make this return int tuples if divisions
ended up integral; I can imagine a situation later where you'd want to
do something with the result of an xyzDiv that's always worked, but now
suddenly doesn't, because some items are of type in. Bit of a pain. Now
I just convert divisor and dividend and divide.