Add tf.setPos, tests
I was writing a test for this, to prove that it returned the name of the
target transform, which I find convenient to chain operations, but then
realized that in Haskel this function would result in type IO (). I'm
going that route for now, because my old semantics don't make sense to
me now; setting an transform's position shouldn't eval to its name.