Update filterBy to work on strings in Python 3
The problem with Python 3 here is that filter needs to be run through
list to get back from the filter object to a useful value. This, in
turn, converts a string to a list of characters. I don't know that this
is a final solution for Python 3, but it seems simple enough, while
being fairly straightforward. Basically, I just check if the input was a
string, and join the list back into a string if so.